+++ examples/cli/config.rs
@@ -1,22 +1,20 @@
-use std::{collections::HashMap, iter::FromIterator};
+use std::{collections::HashMap, iter::FromIterator, sync::LazyLock};
 
-use crate::Loc;
+use crate::loc;
 
 #[derive(Clone)]
 pub struct Config {
     data: HashMap<String, String>,
 }
 
-lazy_static! {
-    // We want to load .env file once per execution and then reuse it
-    static ref CONFIG: Config = Config::load();
-}
+// We want to load .env file once per execution and then reuse it
+static CONFIG: LazyLock<Config> = LazyLock::new(Config::load);
 
 impl Config {
     fn load() -> Self {
         #[allow(deprecated)] // it was undeprecated
         let data = HashMap::from_iter(
-            dotenv::from_path_iter(Loc::env_file().path())
+            dotenv::from_path_iter(loc::env_file().path())
                 .unwrap()
                 .map(Result::unwrap),
         );