+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - [BREAKING] Simplify `Output`: now a struct instead of enum. Use `Output::bytes()` (was `unwrap()`) and `Output::as_string()` (was `unwrap_string()`). Interrupted and killed processes now return `Err(Error::Interrupted)` and `Err(Error::Killed)` instead of `Ok` variants.
 - Add process group support via `SpawnOptions::group`.
 - Add borrowed `IntoIterator` impl for `Env`.
+- Add [`loc!`](https://docs.rs/steward/latest/steward/macro.loc.html) macro for defining project structure/locations.
 
 ### 0.0.7
 - Add [`print`](https://docs.rs/steward/latest/steward/fn.print.html) function.